refactor transaction.py phase 3
