+++ counterparty-core/counterpartycore/lib/transaction_helper/serializer.py
@@ -9,11 +9,9 @@ import decimal
 import hashlib
 import logging
 
-import bitcoin as bitcoinlib
 from bitcoin.bech32 import CBech32Data
 
 from counterpartycore.lib import arc4, backend, config, exceptions, script  # noqa: F401
-from counterpartycore.lib.transaction_helper import p2sh_encoding
 
 logger = logging.getLogger(config.LOGGER_NAME)
 
@@ -339,165 +337,4 @@ def serialise(
                 s += b"\x00"
 
     s += (0).to_bytes(4, byteorder="little")  # LockTime
-    return s
-
-
-def serialise_p2sh_pretx(
-    inputs,
-    source,
-    source_value,
-    data_output,
-    change_output=None,
-    pubkey=None,
-    multisig_pubkeys=None,
-    multisig_pubkeys_required=None,
-):
-    assert data_output  # we don't do this unless there's data
-
-    data_array, data_value = data_output
-
-    s = (1).to_bytes(4, byteorder="little")  # Version
-
-    # Number of inputs.
-    s += var_int(int(len(inputs)))
-
-    # List of Inputs.
-    for i in range(len(inputs)):
-        txin = inputs[i]
-        s += binascii.unhexlify(bytes(txin["txid"], "utf-8"))[::-1]  # TxOutHash
-        s += txin["vout"].to_bytes(4, byteorder="little")  # TxOutIndex
-
-        tx_script = binascii.unhexlify(bytes(txin["script_pub_key"], "utf-8"))
-        s += var_int(int(len(tx_script)))  # Script length
-        s += tx_script  # Script
-        s += b"\xff" * 4  # Sequence
-
-    # Number of outputs.
-    n = len(data_array)
-    if change_output:
-        n += 1
-
-    # encode number of outputs
-    s += var_int(n)
-
-    # P2SH for data encodeded inputs
-    for n, data_chunk in enumerate(data_array):
-        data_chunk = config.PREFIX + data_chunk  # prefix the data_chunk  # noqa: PLW2901
-
-        # get the scripts
-        script_sig, redeem_script, output_script = p2sh_encoding.make_p2sh_encoding_redeemscript(
-            data_chunk, n, pubkey, multisig_pubkeys, multisig_pubkeys_required
-        )
-
-        # if data_value is an array, then every output fee is specified in it
-        if type(data_value) == list:  # noqa: E721
-            s += data_value[n].to_bytes(8, byteorder="little")  # Value
-        else:
-            s += data_value.to_bytes(8, byteorder="little")  # Value
-        s += var_int(int(len(output_script)))  # Script length
-        s += output_script  # Script
-
-    # Change output.
-    if change_output:
-        change_address, change_value = change_output
-        tx_script, witness_script = get_script(change_address)
-
-        s += change_value.to_bytes(8, byteorder="little")  # Value
-        s += var_int(int(len(tx_script)))  # Script length
-        s += tx_script  # Script
-
-    s += (0).to_bytes(4, byteorder="little")  # LockTime
-
-    return s
-
-
-def serialise_p2sh_datatx(
-    txid,
-    source,
-    source_input,
-    destination_outputs,
-    data_output,
-    pubkey=None,
-    multisig_pubkeys=None,
-    multisig_pubkeys_required=None,
-):
-    assert data_output  # we don't do this unless there's data
-
-    txhash = bitcoinlib.core.lx(bitcoinlib.core.b2x(txid))  # reverse txId
-    data_array, value = data_output
-
-    # version
-    s = (1).to_bytes(4, byteorder="little")
-
-    # number of inputs is the length of data_array (+1 if a source_input exists)
-    number_of_inputs = len(data_array)
-    if source_input is not None:
-        number_of_inputs += 1
-    s += var_int(number_of_inputs)
-
-    # Handle a source input here for a P2SH source
-    if source_input is not None:
-        s += binascii.unhexlify(bytes(source_input["txid"], "utf-8"))[::-1]  # TxOutHash
-        s += source_input["vout"].to_bytes(4, byteorder="little")  # TxOutIndex
-
-        # since pubkey is not returned from indexd, add it from bitcoind
-        source_inputs = script.ensure_script_pub_key_for_inputs([source_input])
-        source_input = source_inputs[0]
-        tx_script = binascii.unhexlify(bytes(source_input["script_pub_key"], "utf-8"))
-        s += var_int(int(len(tx_script)))  # Script length
-        s += tx_script  # Script
-        s += b"\xff" * 4  # Sequence
-
-    # list of inputs
-    for n, data_chunk in enumerate(data_array):
-        data_chunk = config.PREFIX + data_chunk  # prefix the data_chunk  # noqa: PLW2901
-
-        # get the scripts
-        script_sig, redeem_script, output_script = p2sh_encoding.make_p2sh_encoding_redeemscript(
-            data_chunk, n, pubkey, multisig_pubkeys, multisig_pubkeys_required
-        )
-        # substituteScript = script_sig + output_script
-
-        s += txhash  # TxOutHash
-        s += (n).to_bytes(4, byteorder="little")  # TxOutIndex (assumes 0-based)
-
-        # s += var_int(len(substituteScript))                      # Script length
-        # s += substituteScript                                    # Script
-
-        s += var_int(len(script_sig))  # + len(output_script))                      # Script length
-        s += script_sig  # Script
-        # s += output_script                                    # Script
-
-        s += b"\xff" * 4  # Sequence
-
-    # number of outputs, always 1 for the opreturn
-    n = 1
-    n += len(destination_outputs)
-
-    # encode output length
-    s += var_int(n)
-
-    # destination outputs
-    for destination, value in destination_outputs:
-        tx_script, witness_script = get_script(destination)
-
-        s += value.to_bytes(8, byteorder="little")  # Value
-        s += var_int(int(len(tx_script)))  # Script length
-        s += tx_script  # Script
-
-    # opreturn to signal P2SH encoding
-    key = arc4.init_arc4(txid)
-    data_chunk = config.PREFIX + b"P2SH"
-    data_chunk = key.encrypt(data_chunk)
-    tx_script = OP_RETURN  # OP_RETURN
-    tx_script += op_push(len(data_chunk))  # Push bytes of data chunk
-    tx_script += data_chunk  # Data
-
-    # add opreturn
-    s += (0).to_bytes(8, byteorder="little")  # Value
-    s += var_int(int(len(tx_script)))  # Script length
-    s += tx_script  # Script
-
-    s += (0).to_bytes(4, byteorder="little")  # LockTime
-
     return s
