+++ counterparty-core/counterpartycore/test/util_test.py
@@ -633,8 +633,6 @@ def run_scenario(scenario):
 
     raw_transactions = []
 
-    transaction.initialise(force=True)
-
     for tx in scenario:
         if tx[0] != "create_next_block":
             mock_protocol_changes = tx[3] if len(tx) == 4 else {}
