+++ counterparty-core/counterpartycore/test/utxolocks_test.py
@@ -27,7 +27,7 @@ def construct_tx(db, source, destination, disable_utxo_locks=False, inputs_set=N
 
 
 def test_utxolocks(server_db):
-    transaction.initialise(force=True)  # reset UTXO_LOCKS
+    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """it shouldn't use the same UTXO"""
     tx1hex = construct_tx(
@@ -50,7 +50,7 @@ def test_utxolocks(server_db):
 
 
 def test_utxolocks_custom_input(server_db):
-    transaction.initialise(force=True)  # reset UTXO_LOCKS
+    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """it should use the same UTXO"""
     inputs_set = [
@@ -91,7 +91,7 @@ def test_utxolocks_custom_input(server_db):
 
 
 def test_disable_utxolocks(server_db):
-    transaction.initialise(force=True)  # reset UTXO_LOCKS
+    transaction.UTXOLocks().init(utxo_locks_max_addresses=10)  # reset UTXO_LOCKS
 
     """with `disable_utxo_locks=True` it should use the same UTXO"""
     tx1hex = construct_tx(
